use float precision

+++ battle/CannonLevelCurve.java
@@ -2,7 +2,6 @@ package common.battle;
 
 import common.util.Data;
 
-import java.util.Arrays;
 import java.util.Map;
 
 public class CannonLevelCurve extends Data  {
@@ -26,7 +25,7 @@ public class CannonLevelCurve extends Data  {
         this.part = part;
     }
 
-    public double applyFormula(int type, int level) {
+    public float applyFormula(int type, int level) {
         if(level <= 0) {
             System.out.println("Warning : Invalid level " + level);
 
@@ -46,32 +45,32 @@ public class CannonLevelCurve extends Data  {
             int max = curve[MAX_VALUE][index];
 
             int minLevel;
-            double v;
+            float v;
 
             if(index == 0) {
                 minLevel = part == PART.CANNON ? 1 : 0;
-                v = min + (max - min) * (level - minLevel) / 9.0;
+                v = min + (max - min) * (level - minLevel) / 9f;
             } else {
                 minLevel = index * 10;
-                v = min + (max - min) * (level - minLevel) / 10.0;
+                v = min + (max - min) * (level - minLevel) / 10f;
             }
 
             switch (part) {
                 case CANNON:
                     switch (type) {
                         case BASE_RANGE:
-                            return (int) v / 4.0;
+                            return (int) v / 4f;
                         case BASE_HEALTH_PERCENTAGE:
-                            return v / 10.0;
+                            return v / 10f;
                         case BASE_HOLY_ATK_SURFACE:
                         case BASE_HOLY_ATK_UNDERGROUND:
-                            return v / 1000.0;
+                            return v / 1000f;
                         default:
                             return v;
                     }
                 case BASE:
                 case DECORATION:
-                    return 1.0 - v / 10000.0;
+                    return 1f - v / 10000f;
             }
         }
