+++ battle/attack/AtkModelEnemy.java
@@ -16,7 +16,7 @@ public class AtkModelEnemy extends AtkModelEntity {
 
 	private final Proc[] cursed;
 
-	protected AtkModelEnemy(EEnemy ent, double d0) {
+	protected AtkModelEnemy(EEnemy ent, float d0) {
 		super(ent, d0, 1);
 		String[] arr = { "KB", "STOP", "SLOW", "WEAK", "WARP", "CURSE", "SNIPER", "SEAL", "POISON", "BOSS",
 				"POIATK", "ARMOR", "SPEED", "DMGCUT", "DMGCAP" };
@@ -42,10 +42,10 @@ public class AtkModelEnemy extends AtkModelEntity {
 				int allow = b.st.data.allow(b, ene);
 
 				if (allow >= 0 || conf.ignore_limit) {
-					int dis = proc.dis == proc.max_dis ? proc.dis : (int) (proc.dis + b.r.nextDouble() * (proc.max_dis - proc.dis + 1));
-					double ep = ent.pos + getDire() * dis;
-					double mula = proc.mult * 0.01;
-					double mult = proc.mult * 0.01;
+					int dis = proc.dis == proc.max_dis ? proc.dis : (int) (proc.dis + b.r.nextFloat() * (proc.max_dis - proc.dis + 1));
+					float ep = ent.pos + getDire() * dis;
+					float mula = proc.mult * 0.01f;
+					float mult = proc.mult * 0.01f;
 
 					if (!conf.fix_buff) {
 						mult *= ((EEnemy) e).mult;
@@ -88,8 +88,8 @@ public class AtkModelEnemy extends AtkModelEntity {
 					lvl = MathUtil.clip(lvl, 1, u.max + u.maxp);
 					lvl *= (100.0 - resist) / 100;
 
-					int dis = proc.dis == proc.max_dis ? proc.dis : (int) (proc.dis + b.r.nextDouble() * (proc.max_dis - proc.dis + 1));
-					double up = ent.pos + getDire() * dis;
+					int dis = proc.dis == proc.max_dis ? proc.dis : (int) (proc.dis + b.r.nextFloat() * (proc.max_dis - proc.dis + 1));
+					float up = ent.pos + getDire() * dis;
 					int minlayer = proc.min_layer, maxlayer = proc.max_layer;
 					if (proc.min_layer == proc.max_layer && proc.min_layer == -1)
 						minlayer = maxlayer = e.layer;
