+++ battle/attack/AttackCanon.java
@@ -7,7 +7,7 @@ import java.util.ArrayList;
 
 public class AttackCanon extends AttackSimple {
 
-	public AttackCanon(Cannon c, int ATK, ArrayList<Trait> tr, int eab, Proc pro, double p0, double p1, int duration) {
+	public AttackCanon(Cannon c, int ATK, ArrayList<Trait> tr, int eab, Proc pro, float p0, float p1, int duration) {
 		super(null, c, ATK, tr, eab, pro, p0, p1, true, null, 9, false, duration);
 		canon = c.id > 2 ? 1 << (c.id - 1) : 1 << c.id;
 		excludeLastEdge = c.id == 6;
