+++ battle/attack/ContAb.java
@@ -11,11 +11,11 @@ public abstract class ContAb extends BattleObj {
 
 	protected final StageBasis sb;
 
-	public double pos;
+	public float pos;
 	public boolean activate = true;
 	public int layer;
 
-	protected ContAb(StageBasis b, double p, int lay) {
+	protected ContAb(StageBasis b, float p, int lay) {
 		sb = b;
 		pos = p;
 		layer = lay;
@@ -23,7 +23,7 @@ public abstract class ContAb extends BattleObj {
 		sb.tlw.sort(Comparator.comparingInt(e -> e.layer));
 	}
 
-	public abstract void draw(FakeGraphics gra, P p, double psiz);
+	public abstract void draw(FakeGraphics gra, P p, float psiz);
 
 	public abstract void update();
